Separate ValuesWrapper and move filter helper
