+++ model/src/helpers.ts
@@ -12,3 +12,18 @@ export const clone = (obj: { [prop: string]: any }) => {
   }
   return obj;
 };
+
+export function filter<T>(
+  obj: T,
+  predicate: (value: any) => boolean
+): Partial<T> {
+  const result = {};
+
+  for (const [key, value] of Object.entries(obj)) {
+    if (value && predicate(value)) {
+      result[key] = value;
+    }
+  }
+
+  return result;
+}
